policy-query-system/requirement.txt
# Core
python-dotenv==1.0.0

# Document ingestion
PyPDF2==3.0.1
python-docx==1.1.0
email-validator==2.1.0

# LLM helpers (Gemini-ready)
google-generativeai==0.7.2

# Text utilities
sentence-transformers==2.7.0
scikit-learn==1.5.1